requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlalchemy
pandas
openpyxl
psycopg2-binary
python-multipart
scikit-learn
umap-learn
hdbscan
